requirements.txt
django
django-extensions
python-dateutil
psychopg2
numpy
south
django-tastypie
